refactor: update import paths to use the correct directory structure and improve store organization
