+++ frontend/src/lib/stores/ScheduleSidebarState.ts
